requirements.txt
neo4j==4.4.11
elasticsearch==7.17.0
python-dotenv==1.0.0
loguru==0.7.2 